Update main.ps1
Script Existence Check: Added checks to confirm the existence of the ask-install.ps1 and ask.ps1 scripts before trying to execute them. This helps prevent a runtime error if these scripts are missing.

Command Execution Success Check: Replaced the use of $LASTEXITCODE with $? to check the success of the last executed command in PowerShell scripts. This is a more reliable method in PowerShell, as $LASTEXITCODE is not always accurate for PowerShell commands. Also, added try-catch blocks around script execution commands to handle potential exceptions and provide meaningful error messages.

Absolute Path Usage: Modified the script to construct absolute paths to the ask-install.ps1 and ask.ps1 scripts. This prevents potential issues if the script is called from a different location, as it ensures the correct scripts are called regardless of the working directory when the script is executed.