Update install-git-and-nodejs.ps1
Check for the existence of scripts: The script assumes the presence of node-install.ps1 and git-install.ps1 in the 'Install' subdirectory. You might want to add checks to confirm that these scripts exist before trying to execute them, which will help to avoid a runtime error.

Check for the success of the command execution: In PowerShell, $LASTEXITCODE is not always a reliable way to check the exit status of the last executed command because it only works for native applications. In the case of PowerShell scripts, you might want to check the $? variable instead, which tells you if the last PowerShell command was successful. You can also make use of try-catch blocks around your commands to handle any exceptions.

Use of relative paths: The script uses relative paths to identify the script files. This is okay as long as the script is always executed from the correct location. If it's ever called from a different location, the relative paths might not work as expected. Using absolute paths could help to avoid this potential problem.